[core] [disjoint] More radian_access functions for box or segment to radian_access.hpp
